Camera improvement
Previously, the camera always looked at the final waypoint at which the pedestrians would end up crossing (if their behavior was to cross). Conversely, now the pedestrian route is generated first, and then the position of the camera is configured based on this path in order to see the pedestrian closer and for more frames.